Using multer to upload and download images and also working with PDFs

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require("../models/product");
 const mongodb = require("mongodb");
+const fileHelper = require("../util/file");
 
 const ObjectId = mongodb.ObjectID;
 const { validationResult } = require("express-validator/check");
@@ -17,12 +18,29 @@ exports.getAddProduct = (req, res, next) => {
   });
 };
 
-exports.postAddProduct = (req, res, nect) => {
+exports.postAddProduct = (req, res, next) => {
   //   products.push({ title: req.body.title });
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+  // console.log(imageUrl);
+  if (!image) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "Add Product",
+      path: "/admin/add-product",
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage: "Attached file is not an image",
+      validationErrors: [],
+      isAuthenticated: req.session.isLoggedIn
+    });
+  }
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     return res.status(422).render("admin/edit-product", {
@@ -32,7 +50,6 @@ exports.postAddProduct = (req, res, nect) => {
       hasError: true,
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description
       },
@@ -41,6 +58,9 @@ exports.postAddProduct = (req, res, nect) => {
       isAuthenticated: req.session.isLoggedIn
     });
   }
+  const imageUrl = image.path;
+  // console.log(imageUrl);
+
   const product = new Product({
     title: title,
     price: price,
@@ -113,7 +133,7 @@ exports.postEditProduct = (req, res, next) => {
   const productId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDescription = req.body.description;
   // console.log(productId);
   const errors = validationResult(req);
@@ -125,7 +145,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDescription,
         _id: productId
@@ -143,7 +162,11 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDescription;
-      product.imageUrl = updatedImageUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
+
       return product.save().then(result => {
         console.log("UPDATED PRODUCT");
         res.redirect("/admin/products");
@@ -176,7 +199,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteproduct = (req, res, next) => {
   const productId = req.body.productId;
-  Product.deleteOne({ _id: productId, userId: req.user._id })
+  Product.findById(productId)
+    .then(product => {
+      if (!product) {
+        return next(new Error("Product not found!!"));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: productId, userId: req.user._id });
+    })
     .then(() => {
       console.log("DESTROYED PRODUCT");
       res.redirect("/admin/products");
